+++ CHANGELOG.md
@@ -12,6 +12,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - PATCH version when backwards compatible bug **fixes** are implemented.
 
 ## [Unreleased]
+### Added
+- rules parameter on DynamicBrcode resource
+- displayDescription parameter on DynamicBrcode resource
 
 ## [2.19.0] - 2024-09-17
 ### Added